Changed the gui image screen thing :(. Also got the model to render and some other shit I can't think of.
There is no purpose that I can possibly think of for that fourth slot. The middle space has to be spices cause they cant fit on the outside and beans in the middle and filter and water..... Yeah, there really was no room for them except the middle. Anyhow, sorry, we can revert that if needed.

I like pies. Raspberry pie tastes like shit. Raspberries taste like shit.